Simplify with new changes
